requirements.txt
fastapi
uvicorn[standard]
pymupdf4llm
python-dotenv
langchain
langchain-community
langchain-openai>=0.1.7
pinecone-client>=3.0.0
langchain-pinecone
numpy
loguru
openai
python-multipart
tiktoken>=0.7.0
tenacity
redis
celery
pydantic-settings
camelot-py[cv]
pdfplumber
opencv-python-headless
pytesseract
Pillow
pandas